# BUG/PERF: merge_asof raising TypeError for various "by" column dtypes